fix: codeStyle, packages names
