+++ test/testofhttphandlers/HttpSubtasksHandlerTest.java
@@ -1,8 +1,7 @@
-package testOfHttpHandlers;
+package testofhttphandlers;
 
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
-
 import httpServer.HttpTaskServer;
 import managerstypes.Managers;
 import managerstypes.TaskManager;
@@ -10,9 +9,9 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import taskstypes.Epic;
+import taskstypes.StatusOfTask;
 import taskstypes.SubTask;
 import taskstypes.Task;
-
 
 import java.io.IOException;
 import java.lang.reflect.Type;
@@ -27,22 +26,29 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-public class HttpEpicsHandlerTest {
+
+public class HttpSubtasksHandlerTest {
+
+    private SubTask subTask;
     private final TaskManager taskManager = Managers.getDefault();
     private final HttpTaskServer httpTaskServer = new HttpTaskServer(taskManager);
     Gson gson = HttpTaskServer.getGson();
 
-    public HttpEpicsHandlerTest() throws IOException {
+    public HttpSubtasksHandlerTest() throws IOException {
     }
 
     @BeforeEach
     void setUp() throws IOException {
         taskManager.removeAllTasks();
-        taskManager.removeAllTasks();
+        taskManager.removeAllSubTasks();
         taskManager.removeAllEpics();
         httpTaskServer.start();
-        Epic epic = new Epic("Эпик 1", "Описание эпика1", 1, new ArrayList<>());
+        Epic epic = new Epic("Имя эпика1", "Описание эпика1");
         taskManager.addEpic(epic);
+        SubTask subTask = new SubTask("Имя подзадачи 1", "Описание подзадачи 1", StatusOfTask.NEW,
+                1, epic.getId(), Duration.ofMinutes(55),
+                LocalDateTime.of(2025, 2, 7, 16, 27));
+        taskManager.addSubTask(subTask);
     }
 
     @AfterEach
@@ -51,8 +57,8 @@ public class HttpEpicsHandlerTest {
     }
 
     @Test
-    void getEpicsTest() throws IOException, InterruptedException {
-        URI uri = URI.create("http://localhost:8080/epics");
+    void getSubtasksTest() throws IOException, InterruptedException {
+        URI uri = URI.create("http://localhost:8080/subtasks");
         HttpRequest request = HttpRequest.newBuilder()
                 .GET()
                 .uri(uri)
@@ -64,46 +70,17 @@ public class HttpEpicsHandlerTest {
 
         assertEquals(200, response.statusCode());
 
-        Type taskType = new TypeToken<ArrayList<Task>>() {}.getType();
-        List<Epic> epics = gson.fromJson(response.body(), taskType);
-
-        assertNotNull(epics);
-        assertEquals(1, epics.size());
-        assertEquals("Эпик 1", epics.getFirst().getTaskName());
-    }
-
-    @Test
-    void getSubtasksByEpicIdTest() throws IOException, InterruptedException {
-        URI uri = URI.create("http://localhost:8080/epics/1/subtasks");
-        SubTask subTask1 = new SubTask("Подзадача 1", "Описание подзадачи1", Duration.ofMinutes(29),
-                LocalDateTime.of(2024, 12, 7, 21, 11), 1);
-        SubTask subTask2 = new SubTask("Подзадача 2", "Описание подзадачи2", Duration.ofMinutes(17),
-                LocalDateTime.of(2025, 2, 7, 17, 27), 1);
-        taskManager.addSubTask(subTask1);
-        taskManager.addSubTask(subTask2);
-        HttpRequest request = HttpRequest.newBuilder()
-                .GET()
-                .uri(uri)
-                .header("Accept", "application/json")
-                .build();
-        HttpClient client = HttpClient.newHttpClient();
-        HttpResponse.BodyHandler<String> handler = HttpResponse.BodyHandlers.ofString();
-        HttpResponse<String> response = client.send(request, handler);
-
-        assertEquals(200, response.statusCode(), "Ожидается код статуса - 200");
-
         Type taskType = new TypeToken<ArrayList<SubTask>>() {}.getType();
         List<SubTask> subTasks = gson.fromJson(response.body(), taskType);
 
         assertNotNull(subTasks);
-        assertEquals(2, subTasks.size());
-        assertEquals("Подзадача 1", subTasks.get(0).getTaskName());
-        assertEquals("Подзадача 2", subTasks.get(1).getTaskName());
+        assertEquals(1, subTasks.size());
+        assertEquals("Имя подзадачи 1", subTasks.getFirst().getTaskName());
     }
 
     @Test
-    void getEpicByIdTest() throws IOException, InterruptedException {
-        URI uri = URI.create("http://localhost:8080/epics/1");
+    void getSubtaskByIdTest() throws IOException, InterruptedException {
+        URI uri = URI.create("http://localhost:8080/subtasks/2");
         HttpRequest request = HttpRequest.newBuilder()
                 .GET()
                 .uri(uri)
@@ -116,17 +93,18 @@ public class HttpEpicsHandlerTest {
         assertEquals(200, response.statusCode());
 
         Type taskType = new TypeToken<Task>() {}.getType();
-        Epic actual = gson.fromJson(response.body(), taskType);
+        SubTask actual = gson.fromJson(response.body(), taskType);
 
-        assertNotNull(actual, "Эпик не возвращается");
-        assertEquals("Эпик 1", actual.getTaskName());
+        assertNotNull(actual);
+        assertEquals("Имя подзадачи 1", actual.getTaskName());
     }
 
     @Test
-    void postEpicTest() throws IOException, InterruptedException {
-        URI uri = URI.create("http://localhost:8080/epics");
-        Epic epic2 = new Epic("Эпик 2", "Описание эпика 2");
-        String json = gson.toJson(epic2);
+    void postSubtaskTest() throws IOException, InterruptedException {
+        URI uri = URI.create("http://localhost:8080/subtasks");
+        SubTask subTask2 = new SubTask("Имя подзадачи 2", "Описание подзадачи 2", Duration.ofMinutes(55),
+                LocalDateTime.of(2025, 2, 7, 17, 27),1);
+        String json = gson.toJson(subTask2);
         HttpRequest request = HttpRequest.newBuilder()
                 .POST(HttpRequest.BodyPublishers.ofString(json))
                 .uri(uri)
@@ -138,13 +116,36 @@ public class HttpEpicsHandlerTest {
         HttpResponse<String> response = client.send(request, handler);
 
         assertEquals(201, response.statusCode());
-        assertEquals("Эпик 2", taskManager.getEpic(2).getTaskName());
-        assertEquals("Описание эпика 2", taskManager.getEpic(2).getDescription());
+        assertEquals("Имя подзадачи 2", taskManager.getSubTask(3).getTaskName());
+        assertEquals("Описание подзадачи 2", taskManager.getSubTask(3).getDescription());
     }
 
     @Test
-    void deleteEpicByIdTest() throws IOException, InterruptedException {
-        URI uri = URI.create("http://localhost:8080/epics/1");
+    void postSubtaskByIdTest() throws IOException, InterruptedException {
+        URI uri = URI.create("http://localhost:8080/subtasks");
+        SubTask subTask2 = new SubTask("Имя подзадачи 2", "Описание подзадачи 2", StatusOfTask.DONE,
+                2, 1, Duration.ofMinutes(55), LocalDateTime.of(2025, 2, 7, 16, 27));
+        String json = gson.toJson(subTask2);
+        HttpRequest request = HttpRequest.newBuilder()
+                .POST(HttpRequest.BodyPublishers.ofString(json))
+                .uri(uri)
+                .header("Accept", "application/json")
+                .header("Content-Type", "application/json")
+                .build();
+        HttpClient client = HttpClient.newHttpClient();
+        HttpResponse.BodyHandler<String> handler = HttpResponse.BodyHandlers.ofString();
+        HttpResponse<String> response = client.send(request, handler);
+
+        assertEquals(201, response.statusCode());
+        assertEquals("Имя подзадачи 2", taskManager.getSubTask(2).getTaskName());
+        assertEquals("Описание подзадачи 2", taskManager.getSubTask(2).getDescription());
+        assertEquals(2, taskManager.getSubTask(2).getId());
+        assertEquals("DONE", taskManager.getSubTask(2).getStatus().toString());
+    }
+
+    @Test
+    void deleteSubtaskByIdTest() throws IOException, InterruptedException {
+        URI uri = URI.create("http://localhost:8080/subtasks/2");
         HttpRequest request = HttpRequest.newBuilder()
                 .DELETE()
                 .uri(uri)
@@ -157,4 +158,3 @@ public class HttpEpicsHandlerTest {
         assertEquals(200, response.statusCode());
     }
 }
-
